requirements.txt
streamlit
google-generativeai
gtts
pydub
speechrecognition
pyaudio
python-dotenv
simpleaudio
